Use decimal instead of float or double for prices in USD so we avoid accuracy issues

+++ Assets/SequenceSDK/Marketplace/CartTests.cs
@@ -64,7 +64,7 @@ namespace Sequence.Marketplace
                     OrderStatus.active, BigInteger.Parse(ChainDictionaries.ChainIdOf[_chain]), possibleAddresses.GetRandomObjectFromArray(), 
                     Random.Range(1, 10000).ToString(), possibleAddresses.GetRandomObjectFromArray(), Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(),
                     Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(), _possibleCurrencyAddresses.GetRandomObjectFromArray(),
-                    Random.Range(1, 10000), priceUSD, Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(),
+                    Random.Range(1, 10000), (decimal)priceUSD, Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(),
                     Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(),
                     Random.Range(1, 19), Random.Range(1, 100), null, DateTime.Now.ToString(CultureInfo.InvariantCulture), DateTime.Now.Add(TimeSpan.FromDays(300)).ToString(CultureInfo.InvariantCulture),
                     DateTime.Now.ToString(CultureInfo.InvariantCulture), DateTime.Now.ToString(CultureInfo.InvariantCulture), DateTime.Now.ToString(CultureInfo.InvariantCulture), DateTime.Now.ToString(CultureInfo.InvariantCulture), ""));
@@ -232,7 +232,7 @@ namespace Sequence.Marketplace
                     OrderStatus.active, BigInteger.Parse(ChainDictionaries.ChainIdOf[_chain]), possibleAddresses.GetRandomObjectFromArray(), 
                     Random.Range(1, 10000).ToString(), possibleAddresses.GetRandomObjectFromArray(), amount.ToString(), Random.Range(1, 10000).ToString(),
                     Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(), _possibleCurrencyAddresses[currencyIndex],
-                    decimals, Random.Range(1f, 10000f), Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(),
+                    decimals, (decimal)Random.Range(1f, 10000f), Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(),
                     Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(), Random.Range(1, 10000).ToString(),
                     Random.Range(1, 19), Random.Range(1, 100), null, DateTime.Now.ToString(CultureInfo.InvariantCulture), DateTime.Now.Add(TimeSpan.FromDays(300)).ToString(CultureInfo.InvariantCulture),
                     DateTime.Now.ToString(CultureInfo.InvariantCulture), DateTime.Now.ToString(CultureInfo.InvariantCulture), DateTime.Now.ToString(CultureInfo.InvariantCulture), DateTime.Now.ToString(CultureInfo.InvariantCulture), ""));
